* Added an Apache::Table analog to the fakerequest utilities.  * Fixed up logging timestamps  * Made error messages for mocks read more clearly.  * Added better error diagnostics for applets to Arrow::AppletRegistry.  * :COMPAT: Made Arrow::Config return nil for non-existant config    members rather than super()ing to Object#method_missing.
--HG--
extra : convert_revision : svn%3Ab4f3b83b-18d6-0310-a2de-c8f40de7a630/trunk%40351

+++ lib/arrow/session/posixlock.rb
@@ -129,7 +129,9 @@ class Arrow::Session::PosixLock < Arrow::Session::Lock
 			path = @lockfile.path
 			@lockfile.close
 			@lockfile = nil
-			File.delete( path.untaint )
+			if File.exist?( path.untaint )
+				File.delete( path.untaint )
+			end
 		end
 	end
